requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-groq
langgraph
faiss-cpu
pytesseract
openai-whisper
pydantic
python-dotenv
requests
pillow
numpy
sentence-transformers
sympy
streamlit-option-menu
streamlit-lottie
pandas
python-multipart
